requirements.txt
tensorflow
pandas
numpy
scikit-learn
opencv-python
pytesseract
fuzzywuzzy
python-Levenshtein
matplotlib
streamlit
tqdm
Pillow
